feat(phase-19/11): expand llm-observability-dashboard into multi-file TypeScript project
